+++ Servicio.RetazoMarket.DataAccess/Entities/ProductoMaterialEntity.cs
@@ -4,7 +4,7 @@ namespace Servicio.RetazoMarket.DataAccess.Entities
     {
         public int id_producto { get; set; }
         public int id_material { get; set; }
-        public int cantidad_req { get; set; }
+        public decimal cantidad_req { get; set; }
         public string es_personalizable { get; set; } = null!;
 
         public ProductoEntity Producto { get; set; } = null!;
