+++ Servicio.RetazoMarket.DataManagment/Models/MaterialFiltroDataModel.cs
@@ -5,8 +5,8 @@ namespace Servicio.RetazoMarket.DataManagment.Models
         public string? nombre { get; set; }
         public int? id_categoria { get; set; }
         public string? estado { get; set; }
-        public int? stockMinimo { get; set; }
-        public int? stockMaximo { get; set; }
+        public decimal? stockMinimo { get; set; }
+        public decimal? stockMaximo { get; set; }
         public int PageNumber { get; set; } = 1;
         public int PageSize { get; set; } = 10;
     }
