+++ Servicio.RetazoMarket.DataManagment/Models/MovimientoMaterialDataModel.cs
@@ -5,7 +5,7 @@ namespace Servicio.RetazoMarket.DataManagment.Models
         public int id_movimiento { get; set; }
         public int id_material { get; set; }
         public string tipo_movimiento { get; set; } = null!;
-        public int cantidad { get; set; }
+        public decimal cantidad { get; set; }
         public DateTime fecha_mov { get; set; }
         public string motivo_mov { get; set; } = null!;
         public string? metodo_pago { get; set; }
@@ -17,6 +17,6 @@ namespace Servicio.RetazoMarket.DataManagment.Models
         public int id_material { get; set; }
         public string mat_nombre { get; set; } = null!;
         public string unidad_medida { get; set; } = null!;
-        public int stock_actual { get; set; }
+        public decimal stock_actual { get; set; }
     }
 }
